Move FlowHandle implementations back into core so that RPC clients can use them.
