+++ core/src/main/kotlin/net/corda/core/messaging/FlowHandle.kt
@@ -2,6 +2,7 @@ package net.corda.core.messaging
 
 import com.google.common.util.concurrent.ListenableFuture
 import net.corda.core.flows.StateMachineRunId
+import net.corda.core.serialization.CordaSerializable
 import rx.Observable
 
 /**
@@ -34,4 +35,38 @@ interface FlowProgressHandle<A> : FlowHandle<A> {
      * Note that it won't really close if one subscribes on progress [Observable], but then forgets to unsubscribe.
      */
     override fun close()
+}
+
+
+@CordaSerializable
+data class FlowHandleImpl<A>(
+    override val id: StateMachineRunId,
+    override val returnValue: ListenableFuture<A>) : FlowHandle<A> {
+
+     // Remember to add @Throws to FlowHandle.close() if this throws an exception.
+    override fun close() {
+        returnValue.cancel(false)
+    }
+}
+
+@CordaSerializable
+data class FlowProgressHandleImpl<A>(
+    override val id: StateMachineRunId,
+    override val returnValue: ListenableFuture<A>,
+    override val progress: Observable<String>) : FlowProgressHandle<A> {
+
+    // Remember to add @Throws to FlowProgressHandle.close() if this throws an exception.
+    override fun close() {
+        progress.notUsed()
+        returnValue.cancel(false)
+    }
+}
+
+// Private copy of the version in client:rpc.
+private fun <T> Observable<T>.notUsed() {
+    try {
+        this.subscribe({}, {}).unsubscribe()
+    } catch (e: Exception) {
+        // Swallow any other exceptions as well.
+    }
 }
